+++ ceammc/gui/plugins/gen_popup_menu.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-cat ceammc_popup_menu_vanilla.tcl \
-    ceammc_popup_menu_ceammc.tcl \
-    ceammc_popup_menu_extra.tcl\
-    > ceammc_popup_menu_tree.tcl
+cat "@CMAKE_CURRENT_SOURCE_DIR@/ceammc_popup_menu_vanilla.tcl" \
+    "@CMAKE_CURRENT_SOURCE_DIR@/ceammc_popup_menu_ceammc.tcl" \
+    "@CMAKE_CURRENT_BINARY_DIR@/ceammc_popup_menu_extra.tcl" \
+    > "@CMAKE_CURRENT_BINARY_DIR@/ceammc_popup_menu_tree.tcl"